# Conflicts:
#	src/main/java/com/example/demo/security/config/CustomSecurityConfig.java
#	src/main/java/com/example/demo/security/filter/JWTCheckFilter.java

+++ src/main/java/com/example/demo/manager/repository/StoreOwnerRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StoreOwnerRepository extends JpaRepository<StoreOwner, Integer> {
+public interface StoreOwnerRepository extends JpaRepository<StoreOwner, Long> {
 
     // 점주 ID로 조회하는 메서드 추가
     Optional<StoreOwner> findById(String id);
